requirements.txt
torchmetrics
mamba_ssm
omegaconf
loguru
timm